requirements.txt
telebot~=0.0.5
python-dotenv~=1.0.1
redis~=5.0.1
requests~=2.31.0
beautifulsoup4~=4.12.3